Repackage all of the annotation processors into 'processor' subpackages.
