+++ factory/src/main/java/com/google/auto/factory/processor/ImplemetationMethodDescriptor.java
@@ -13,10 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.auto.factory;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Preconditions.checkState;
+package com.google.auto.factory.processor;
 
 import java.util.Set;
 
@@ -26,37 +23,19 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 
-/**
- * A value object representing a factory method to be generated.
- *
- * @author Gregory Kick
- */
-final class FactoryMethodDescriptor {
-  private final AutoFactoryDeclaration declaration;
+final class ImplemetationMethodDescriptor {
   private final String factoryName;
   private final String name;
   private final String returnType;
   private final boolean publicMethod;
-  private final boolean override;
   private final ImmutableSet<Parameter> passedParameters;
-  private final ImmutableSet<Parameter> providedParameters;
-  private final ImmutableSet<Parameter> creationParameters;
 
-  private FactoryMethodDescriptor(Builder builder) {
-    this.declaration = builder.declaration;
+  private ImplemetationMethodDescriptor(Builder builder) {
     this.factoryName = builder.factoryName.get();
     this.name = builder.name.get();
     this.returnType = builder.returnType.get();
     this.publicMethod = builder.publicMethod;
-    this.override = builder.override;
     this.passedParameters = ImmutableSet.copyOf(builder.passedParameters);
-    this.providedParameters = ImmutableSet.copyOf(builder.providedParameters);
-    this.creationParameters = ImmutableSet.copyOf(builder.creationParameters);
-    checkState(creationParameters.equals(Sets.union(passedParameters, providedParameters)));
-  }
-
-  AutoFactoryDeclaration declaration() {
-    return declaration;
   }
 
   String factoryName() {
@@ -71,24 +50,12 @@ final class FactoryMethodDescriptor {
     return returnType;
   }
 
-  public boolean publicMethod() {
+  boolean publicMethod() {
     return publicMethod;
-  }
-
-  public boolean override() {
-    return override;
   }
 
   ImmutableSet<Parameter> passedParameters() {
     return passedParameters;
-  }
-
-  ImmutableSet<Parameter> providedParameters() {
-    return providedParameters;
-  }
-
-  ImmutableSet<Parameter> creationParameters() {
-    return creationParameters;
   }
 
   @Override
@@ -97,25 +64,17 @@ final class FactoryMethodDescriptor {
         .add("factoryName", factoryName)
         .add("name", name)
         .add("returnType", returnType)
-        .add("passed", passedParameters)
-        .add("provided", providedParameters)
+        .add("publicMethod", publicMethod)
+        .add("passedParameters", passedParameters)
         .toString();
   }
 
   static final class Builder {
-    private final AutoFactoryDeclaration declaration;
     private Optional<String> factoryName = Optional.absent();
     private Optional<String> name = Optional.absent();
     private Optional<String> returnType = Optional.absent();
     private boolean publicMethod = false;
-    private boolean override = false;
     private final Set<Parameter> passedParameters = Sets.newLinkedHashSet();
-    private final Set<Parameter> providedParameters = Sets.newLinkedHashSet();
-    private final Set<Parameter> creationParameters = Sets.newLinkedHashSet();
-
-    Builder(AutoFactoryDeclaration declaration) {
-      this.declaration = checkNotNull(declaration);
-    }
 
     Builder factoryName(String factoryName) {
       this.factoryName = Optional.of(factoryName);
@@ -141,28 +100,13 @@ final class FactoryMethodDescriptor {
       return this;
     }
 
-    Builder override() {
-      this.override = true;
-      return this;
-    }
-
     Builder passedParameters(Iterable<Parameter> passedParameters) {
       Iterables.addAll(this.passedParameters, passedParameters);
       return this;
     }
 
-    Builder providedParameters(Iterable<Parameter> providedParameters) {
-      Iterables.addAll(this.providedParameters, providedParameters);
-      return this;
-    }
-
-    Builder creationParameters(Iterable<Parameter> creationParameters) {
-      Iterables.addAll(this.creationParameters, creationParameters);
-      return this;
-    }
-
-    FactoryMethodDescriptor build() {
-      return new FactoryMethodDescriptor(this);
+    ImplemetationMethodDescriptor build() {
+      return new ImplemetationMethodDescriptor(this);
     }
   }
 }
