+++ factory/src/main/java/com/google/auto/factory/processor/Key.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.auto.factory;
+package com.google.auto.factory.processor;
 
 import com.google.common.base.Objects;
 import com.google.common.base.Optional;
