+++ factory/src/test/java/com/google/auto/factory/processor/AutoFactoryProcessorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.auto.factory;
+package com.google.auto.factory.processor;
 
 import static com.google.testing.compile.JavaSourceSubjectFactory.javaSource;
 import static com.google.testing.compile.JavaSourcesSubjectFactory.javaSources;
@@ -23,6 +23,7 @@ import javax.tools.JavaFileObject;
 
 import org.junit.Test;
 
+import com.google.auto.factory.processor.AutoFactoryProcessor;
 import com.google.common.collect.ImmutableSet;
 import com.google.testing.compile.JavaFileObjects;
