+++ service/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-com.google.auto.service.AutoServiceProcessor
+com.google.auto.service.processor.AutoServiceProcessor
