+++ value/src/main/java/com/google/auto/value/processor/AbstractMethodLister.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.auto.value;
+package com.google.auto.value.processor;
 
 import java.io.IOException;
 import java.io.InputStream;
